refactor: merge service classes
