+++ src/main/java/lotto/controller/PurchasingLottoController.java
@@ -3,7 +3,7 @@ package lotto.controller;
 import lotto.domain.LottoTickets;
 import lotto.domain.Money;
 import lotto.domain.factory.LottoTicketsFactory;
-import lotto.service.LottoTicketService;
+import lotto.service.LottoGameService;
 import spark.Route;
 
 import java.util.*;
@@ -30,7 +30,7 @@ public class PurchasingLottoController {
         model.put("amountOfCustom", inputCustoms.size());
         model.put("amountOfAuto", (money.getMoney() / MONEY_OFFSET) - inputCustoms.size());
 
-        LottoTicketService.addLottoTicket(lottoTickets);
+        LottoGameService.addLottoTicket(lottoTickets);
 
         return render(model, "purchased_tickets.html");
     };
